Use Floats Instead of Doubles

+++ dsp-stuff-win32/src/audio/audio_util.hpp
@@ -7,13 +7,19 @@
 
 const unsigned scale = (1 << 23) - 1;
 
-inline unsigned scaleSignal(double sig) {
-    double f = ((sig * 0.5) + 0.5) * scale;
+// inline unsigned scaleSignal(double sig) {
+//     double f = ((sig * 0.5) + 0.5) * scale;
+//     unsigned u = (unsigned)f << 8;
+//     return u;
+// }
+
+inline unsigned scaleSignal(float sig) {
+    float f = ((sig * 0.5f) + 0.5f) * scale;
     unsigned u = (unsigned)f << 8;
     return u;
 }
 
-inline unsigned mstosamps(double ms) {
+inline unsigned mstosamps(float ms) {
     return (unsigned)(ms * samplesPerMs);
 }
